requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
asyncpg==0.30.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
aiohttp==3.11.13
alembic==1.14.1
gunicorn==23.0.0
pydantic==2.10.6
email-validator==2.2.0
yarl==1.18.3
asyncio==3.4.3
redis==5.2.1
requests~=2.32.3
pytest==8.0.0
pytest-cov
pytest-asyncio
aiogram~=3.18.0
numpy~=2.2.3
httpx~=0.28.1
prometheus_client==0.21.1
pytz~=2025.1
boto3~=1.37.4
pillow~=11.1.0
botocore~=1.37.4
python-multipart~=0.0.20